Significant doc improvement for Queries (#161). Perhaps more after review.
git-svn-id: file:///home/svn/omero/trunk@727 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/server/src/ome/services/util/CountCollector.java
@@ -116,7 +116,7 @@ public class CountCollector extends ProxySafeFilter {
         
     }
     
-    public void addCounts(String field, Long id, Integer count) 
+    public void addCounts(String field, Long id, Long count) 
     {
         
         int idx = Arrays.binarySearch(fields,field);
